# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API Quản trị Nhân sự HUIT
  description: Tài liệu RESTful API cho hệ thống Quản trị nhân sự HUIT (ERP Backend) do Hóa Coder phát triển.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Dev server kết nối Database Local/Docker
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Employee:
      type: object
      properties:
        hoten:
          type: string
        email:
          type: string
        chucvu:
          type: string
        luong:
          type: number
        maphg:
          type: integer
        ngaysinh:
          type: string
          format: date
        gioitinh:
          type: string
        diachinhan:
          type: string
...
  responses:
    SuccessCreate:
      description: Thêm mới thành công
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
    SuccessUpdate:
      description: Cập nhật thành công
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      summary: Lấy danh sách toàn bộ nhân viên (Có phân trang)
      tags: [Employees]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Trả về danh sách
    post:
      summary: HR/Admin tạo mới nhân sự
      tags: [Employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          $ref: '#/components/responses/SuccessCreate'
  /employees/{id}:
    get:
      summary: Xem chi tiết hồ sơ 1 nhân viên
      tags: [Employees]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trả về thông tin NV
    put:
      summary: Cập nhật thông tin nhân viên
      tags: [Employees]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          $ref: '#/components/responses/SuccessUpdate'

  # =========================================================================
  # PHÒNG BAN
  # =========================================================================
...
  /projects/employee/{id}:
    get:
      summary: Lấy danh sách các dự án mà 1 nhân viên tham gia
      tags: [Projects]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: Mã nhân viên
      responses:
        '200':
          description: Danh sách dự án cá nhân
...