# routers/employee.ts
import express from "express";
const router = express.Router();
import sql from "msnodesqlv8";
import withUserConnection from "../middleware/authMiddleware";
import employeeController from "../controllers/employeeController";
import { sql as globalSql } from "../config/db";

// ⚠️ ROUTES SPECIFIC PHẢI TRƯỚC GENERIC ROUTES (/:id)

// 1️⃣ SPECIFIC ROUTES (profile, my-projects, coworkers, update-info)
// GET /api/employees/profile/:manv - Xem profile cá nhân (dùng dynamic connection)
router.get("/profile/:manv", withUserConnection, (req, res) => {
  const query = `
    SELECT nv.MANV, nv.HOTEN, nv.LUONG, nv.CHUCVU, nv.EMAIL, pb.TENPB, hs.SO_CCCD
    FROM NHAN_VIEN nv
    LEFT JOIN PHONG_BAN pb ON nv.MAPHG = pb.MAPHG
    LEFT JOIN HO_SO_BM hs ON nv.MANV = hs.MANV
    WHERE nv.MANV = ?`;

  sql.query(req.userConnectionString, query, [req.params.manv], (err, rows) => {
    if (err) {
      console.error("Lỗi quyền hạn:", err.message);
      return res
        .status(403)
        .json({ error: "Bạn không có quyền xem dữ liệu này (SQL Denied)" });
    }
    rows.length > 0
      ? res.json(rows[0])
      : res.status(404).json({ message: "404" });
  });
});

// GET /api/employees/my-projects/:manv - Xem dự án của tôi (dùng dynamic connection)
router.get("/my-projects/:manv", withUserConnection, (req, res) => {
  const query = `SELECT da.TENDA, pc.THOIGIAN FROM PHANCONG pc JOIN DUAN da ON pc.MADA = da.MADA WHERE pc.MANV = ?`;
  sql.query(req.userConnectionString, query, [req.params.manv], (err, rows) => {
    if (err)
      return res.status(500).json({ error: "Lỗi truy vấn hoặc quyền hạn!" });
    res.json(rows);
  });
});

// GET /api/employees/coworkers/:maphg - Xem đồng nghiệp cùng phòng (dùng dynamic connection)
router.get("/coworkers/:maphg", withUserConnection, (req, res) => {
  sql.query(
    req.userConnectionString,
    "SELECT MANV, HOTEN, CHUCVU FROM NHAN_VIEN WHERE MAPHG = ?",
    [req.params.maphg],
    (err, rows) => {
      if (err) return res.status(403).json({ error: "Access Denied" });
      res.json(rows);
    },
  );
});

// PUT /api/employees/update-info - Cập nhật thông tin cá nhân (dùng SA connection từ global pool)
router.put("/update-info", async (req, res) => {
  try {
    const { manv, email } = req.body;
    const request = new globalSql.Request();
    await request
      .input("MaNV", globalSql.NVarChar, manv)
      .input("Email", globalSql.NVarChar, email)
      .query("UPDATE NHAN_VIEN SET EMAIL = @Email WHERE MANV = @MaNV");
    res.json({ message: "Cập nhật thành công" });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// 2️⃣ GENERIC ROUTES (list, detail, create, update, delete)
// GET /api/employees - Lấy danh sách nhân viên
router.get("/", employeeController.getAllEmployees);

// POST /api/employees - Thêm nhân viên mới
router.post("/", employeeController.createEmployee);

// GET /api/employees/:id - Xem chi tiết 1 nhân viên
router.get("/:id", employeeController.getEmployeeById);

// PUT /api/employees/:id - Cập nhật thông tin nhân viên
router.put("/:id", employeeController.updateEmployee);

// DELETE /api/employees/:id - Xóa/Khóa nhân viên
router.delete("/:id", employeeController.deleteEmployee);

export default router;